+++ test/Telegram.Bot.Tests.Integ/Games/GamesExceptionTests.cs
@@ -1,0 +1,83 @@
+﻿using System.Threading.Tasks;
+using Telegram.Bot.Exceptions;
+using Telegram.Bot.Tests.Integ.Framework;
+using Xunit;
+
+namespace Telegram.Bot.Tests.Integ.Games
+{
+    [Collection(Constants.TestCollections.Games)]
+    [TestCaseOrderer(Constants.TestCaseOrderer2, Constants.AssemblyName)]
+    public class GamesExceptionTests : IClassFixture<GamesFixture>
+    {
+        private ITelegramBotClient BotClient => _fixture.BotClient;
+
+        private readonly TestsFixture _fixture;
+
+        private readonly GamesFixture _classFixture;
+
+        public GamesExceptionTests(TestsFixture fixture, GamesFixture classFixture)
+        {
+            _fixture = fixture;
+            _classFixture = classFixture;
+        }
+
+        [OrderedFact(DisplayName = FactTitles.ShouldThrowInvalidGameShortNameException)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendGame)]
+        public async Task Should_Throw_InvalidGameShortNameException()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowInvalidGameShortNameException);
+
+            InvalidGameShortNameException e = await Assert.ThrowsAsync<InvalidGameShortNameException>(() =>
+                BotClient.SendGameAsync(
+                    chatId: _fixture.SupergroupChat.Id,
+                    gameShortName: "my game"
+                )
+            );
+
+            Assert.Equal("game_short_name", e.Parameter);
+        }
+
+        [OrderedFact(DisplayName = FactTitles.ShouldThrowInvalidGameShortNameException2)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendGame)]
+        public async Task Should_Throw_InvalidGameShortNameException_2()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowInvalidGameShortNameException2);
+
+            InvalidGameShortNameException e = await Assert.ThrowsAsync<InvalidGameShortNameException>(() =>
+                BotClient.SendGameAsync(
+                    chatId: _fixture.SupergroupChat.Id,
+                    gameShortName: string.Empty
+                )
+            );
+
+            Assert.Equal("game_short_name", e.Parameter);
+        }
+        
+        [OrderedFact(DisplayName = FactTitles.ShouldThrowInvalidGameShortNameException3)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendGame)]
+        public async Task Should_Throw_InvalidGameShortNameException_3()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowInvalidGameShortNameException3);
+
+            InvalidGameShortNameException e = await Assert.ThrowsAsync<InvalidGameShortNameException>(() =>
+                BotClient.SendGameAsync(
+                    chatId: _fixture.SupergroupChat.Id,
+                    gameShortName: "non_existing_game"
+                )
+            );
+
+            Assert.Equal("game_short_name", e.Parameter);
+        }
+
+        private static class FactTitles
+        {
+            public const string ShouldThrowInvalidGameShortNameException = "Should throw InvalidGameShortNameException";
+
+            public const string ShouldThrowInvalidGameShortNameException2 =
+                "Should throw InvalidGameShortNameException for empty name";
+
+            public const string ShouldThrowInvalidGameShortNameException3 =
+                "Should throw InvalidGameShortNameException for non-existing game";
+        }
+    }
+}